requirements.txt
pip==24.1.2
databricks-cli==0.18.0
azure-storage-blob==12.20.0
azure-identity==1.17.1